requirements.txt
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
Django==2.2
pycparser==2.19
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
